+++ index.js
@@ -18,6 +18,7 @@ app.use('/', require(path.join(__dirname, "/routes/pages/root.js")));
 app.use('/api/components', require(path.join(__dirname, "/routes/api/components.js")));
 app.use('/api/auth', require(path.join(__dirname, "/routes/api/authentication.js")));
 app.use('/api/cars', require(path.join(__dirname, "/routes/api/create-cars-data.js")));
+app.use('/api/subscribe', require(path.join(__dirname, "/routes/api/email-subscribe.js")));
 
 
 app.listen(port, () => {